Initial values were adjusted.

+++ src/ukf.h
@@ -32,7 +32,7 @@ public:
   MatrixXd Xsig_pred_;
 
   ///* time when the state is true, in us
-  long long previous_timestamp_;
+  double previous_timestamp_;
 
   ///* Process noise standard deviation longitudinal acceleration in m/s^2
   double std_a_;
@@ -65,7 +65,7 @@ public:
   int n_aug_;
 
   ///* Sigma point spreading parameter
-  double lambda_;
+  int lambda_;
 
   ///* the current NIS for radar
   double NIS_radar_;
